refactor: spliting components into dedicated folders
