+++ components/items/components/items-in-table.tsx
@@ -1,14 +1,15 @@
+// COMPONENTS
 import {
   Table,
   TableBody,
   TableCaption,
-  TableCell,
   TableHead,
   TableHeader,
   TableRow,
 } from "@/components/ui/table"
+import { ItemRow } from "@/components/item/item-row"
+// TYPES
 import { IQueryItem } from "@/types/firestore"
-import { ItemStatus } from "./item-status"
 
 export const ItemsInTable = ({ items }: { items : IQueryItem[] | null }) => {
   if (!items?.length) return null;
@@ -19,7 +20,7 @@ export const ItemsInTable = ({ items }: { items : IQueryItem[] | null }) => {
       <ItemsTableHeader />
       <TableBody>
         {items.map((item) => (
-          <ItemsTableRow key={item.id} item={item} />
+          <ItemRow key={item.id} item={item} />
         ))}
       </TableBody>
     </Table>
@@ -36,16 +37,3 @@ export const ItemsTableHeader = () => (
     </TableRow>
   </TableHeader>
 )
-
-export const ItemsTableRow = ({ item }: { item : IQueryItem | null }) => {
-  if (!item) return null;
-  
-  return (
-    <TableRow>
-      <TableCell className="font-medium">{item.id}</TableCell>
-      <TableCell>{<ItemStatus status="INITIAL"/>}</TableCell>
-      <TableCell>{item.additionalInfo}</TableCell>
-      <TableCell className="text-right">$250.00</TableCell>
-    </TableRow>
-  )
-}
